Add Condition status is enable then work and add product select in config

+++ custom/plugins/ICTCustomField/src/ICTCustomField.php
@@ -22,9 +22,9 @@ class ICTCustomField extends Plugin
     {
         parent::uninstall($uninstallContext);
 
-//        if ($uninstallContext->keepUserData()) {
-//            return;
-//        }
+       if ($uninstallContext->keepUserData()) {
+           return;
+       }
 
         $this->getICTCustomFieldsInstaller()->deleteRelations($uninstallContext->getContext());
         // Remove or deactivate the data created by the plugin
